// common/randomizer/enums/PrimeItem.ts
export enum PrimeItem {
  MISSILE_EXPANSION = 'Missile Expansion',
  ENERGY_TANK = 'Energy Tank',
  POWER_BOMB_EXPANSION = 'Power Bomb Expansion',
  MISSILE_LAUNCHER = 'Missile Launcher',
  WAVE_BEAM = 'Wave Beam',
  ICE_BEAM = 'Ice Beam',
  PLASMA_BEAM = 'Plasma Beam',
  CHARGE_BEAM = 'Charge Beam',
  SPACE_JUMP_BOOTS = 'Space Jump Boots',
  SUPER_MISSILE = 'Super Missile',
  WAVEBUSTER = 'Wavebuster',
  ICE_SPREADER = 'Ice Spreader',
  FLAMETHROWER = 'Flamethrower',
  GRAPPLE_BEAM = 'Grapple Beam',
  MORPH_BALL = 'Morph Ball',
  BOOST_BALL = 'Boost Ball',
  SPIDER_BALL = 'Spider Ball',
  MORPH_BALL_BOMB = 'Morph Ball Bomb',
  POWER_BOMB = 'Power Bomb',
  VARIA_SUIT = 'Varia Suit',
  GRAVITY_SUIT = 'Gravity Suit',
  PHAZON_SUIT = 'Phazon Suit',
  THERMAL_VISOR = 'Thermal Visor',
  XRAY_VISOR = 'X-Ray Visor',
  ARTIFACT_OF_TRUTH = 'Artifact of Truth',
  ARTIFACT_OF_STRENGTH = 'Artifact of Strength',
  ARTIFACT_OF_ELDER = 'Artifact of Elder',
  ARTIFACT_OF_WILD = 'Artifact of Wild',
  ARTIFACT_OF_LIFEGIVER = 'Artifact of Lifegiver',
  ARTIFACT_OF_WARRIOR = 'Artifact of Warrior',
  ARTIFACT_OF_CHOZO = 'Artifact of Chozo',
  ARTIFACT_OF_NATURE = 'Artifact of Nature',
  ARTIFACT_OF_SUN = 'Artifact of Sun',
  ARTIFACT_OF_WORLD = 'Artifact of World',
  ARTIFACT_OF_SPIRIT = 'Artifact of Spirit',
  ARTIFACT_OF_NEWBORN = 'Artifact of Newborn'
}